requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
psycopg2-binary
pgvector
gunicorn
SQLAlchemy